mejoro funcionalidad para el instalador

+++ MediaCenter/deploy.sh
@@ -1,22 +1,27 @@
-dpkg --configure -a
-
 source ./conf.sh
 
-source ./Config/ConfigVBoxLinuxAdditions.sh
+if [ "$1" = "1" ];then
+	dpkg --configure -a
 
-source ./Config/ConfigIP.sh
+	source ./Config/ConfigVBoxLinuxAdditions.sh
 
-#source ./Config/ConfigDocker.sh
+	source ./Config/ConfigIP.sh
+	
+	halt --reboot
+fi
+if [ "$1" = "2" ];then
+	source ./Config/ConfigDocker.sh
 
-#source ./Config/ConfigGroups.sh
+	source ./Config/ConfigGroups.sh
 
-#source ./Config/ConfigFolders.sh
+	source ./Config/ConfigFolders.sh
 
-#source ./Config/ConfigTransmission.sh
+	source ./Config/ConfigTransmission.sh
 
-#source ./Config/ConfigCouchpotato.sh
+	source ./Config/ConfigCouchpotato.sh
 
-#source ./Config/ConfigSickrage.sh
+	source ./Config/ConfigSickrage.sh
 
+	source ./Config/ConfigShipYard.sh
 
-halt --reboot
+fi
